+++ json-to-messages/javascript/src/cucumber-ruby/JSONSchema.ts
@@ -1,4 +1,4 @@
-import { IMatch, IResult, ITag } from '../cucumber-generic/JSONSchema'
+import { IMatch, IResult, ITag } from '../types'
 
 export interface IElement {
   line: number
