+++ json-to-messages/javascript/test/jsonToMessagesTest.ts
@@ -1,12 +1,7 @@
 import { Readable } from 'stream'
 import { jsonToMessages } from '../src'
-import {
-  messages,
-  version as messagesVersion,
-} from '@cucumber/messages'
-import {
-  NdjsonToMessageStream,
-} from '@cucumber/messages/dist/src/stream'
+import { messages, version as messagesVersion } from '@cucumber/messages'
+import { NdjsonToMessageStream } from '@cucumber/messages/dist/src/stream'
 import assert from 'assert'
 import { version } from '../package.json'
